# refactor: optimize MiniBlockRepIndex decode by decoding from bytes directly