`MiniBlockRepIndex::decode` used to be an allocation-heavy operation, creating numerous `Vec<Vec<u8>>` instances in memory only to drop them after parsing. The idea behind this PR is straightforward: decode the rep index directly from bytes without introducing any additional allocations.

## Before

```shell
Benchmarking V2_1 Single Random Take FileReader(1048577 file size, 1024 batches, 1000 rows per take)
Benchmarking V2_1 Single Random Take FileReader(1048577 file size, 1024 batches, 1000 rows per take): Warming up for 3.0000 s

Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 16.7s, or reduce sample count to 20.
Benchmarking V2_1 Single Random Take FileReader(1048577 file size, 1024 batches, 1000 rows per take): Collecting 100 samples in estimated 16.699 s (100 iterations)
Benchmarking V2_1 Single Random Take FileReader(1048577 file size, 1024 batches, 1000 rows per take): Analyzing
V2_1 Single Random Take FileReader(1048577 file size, 1024 batches, 1000 rows per take)
                        time:   [162.71 ms 163.23 ms 163.74 ms]
                        change: [+1.7266% +2.6730% +3.6496%] (p = 0.00 < 0.10)
                        Performance has regressed.
```

The `MiniBlockScheduler::initiate` only takes 45.3% of whole time.

<img width="3224" height="1126" alt="image" src="https://github.com/user-attachments/assets/d7e694db-1d2c-4fd7-b797-b7df614a8b2b" />


## After

```shell
Benchmarking V2_1 Single Random Take FileReader(1048577 file size, 1024 batches, 1000 rows per take)
Benchmarking V2_1 Single Random Take FileReader(1048577 file size, 1024 batches, 1000 rows per take): Warming up for 3.0000 s

Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 10.0s, or reduce sample count to 40.

Benchmarking V2_1 Single Random Take FileReader(1048577 file size, 1024 batches, 1000 rows per take): Collecting 100 samples in estimated 10.034 s (100 iterations)
Benchmarking V2_1 Single Random Take FileReader(1048577 file size, 1024 batches, 1000 rows per take): Analyzing
V2_1 Single Random Take FileReader(1048577 file size, 1024 batches, 1000 rows per take)
                        time:   [92.341 ms 93.659 ms 94.998 ms]
                        change: [-43.446% -42.621% -41.718%] (p = 0.00 < 0.10)
                        Performance has improved.
```

The `MiniBlockScheduler::initiate` only takes 21.2% of whole time.

<img width="3622" height="1334" alt="image" src="https://github.com/user-attachments/assets/195c1e25-cf5d-4df8-a85f-c2c556413936" />

## Summary

This PR will make lance `2.1` 50% faster than before.

However we are still slower than 2.0, I will continute working on this.

```shell
Benchmarking V2_0 Single Random Take FileReader(1048577 file size, 1024 batches, 1000 rows per take): Collecting 10 samples in V2_0 Single Random Take FileReader(1048577 file size, 1024 batches, 1000 rows per take)
                        time:   [31.472 ms 31.659 ms 31.938 ms]
                        change: [-1.5902% -0.5624% +0.4912%] (p = 0.33 > 0.10)
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe

Benchmarking V2_1 Single Random Take FileReader(1048577 file size, 1024 batches, 1000 rows per take): Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.4s or enable flat sampling.
Benchmarking V2_1 Single Random Take FileReader(1048577 file size, 1024 batches, 1000 rows per take): Collecting 10 samples in V2_1 Single Random Take FileReader(1048577 file size, 1024 batches, 1000 rows per take)
                        time:   [87.263 ms 90.066 ms 94.046 ms]
                        change: [-6.6512% -2.6566% +2.7044%] (p = 0.32 > 0.10)
                        No change in performance detected.
```